file-manager/requirements.txt
# =================================================================
# FASTAPI Y SERVIDOR
# =================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# =================================================================
# GOOGLE DRIVE API
# =================================================================
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# =================================================================
# UTILIDADES
# =================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0